Added loadbalancer to charm

+++ src/tools/CollectorCharm.hpp
@@ -46,7 +46,7 @@ public:
 
                 instance.measure_ctrs[i] = std::chrono::steady_clock::time_point(std::chrono::milliseconds((long)arr[i+5]));
                 if(i==CTR_WALL)
-                    std::cout<<  "recv measure_ctr "<<(long)arr[i+5] << std::endl;
+                    std::cout<<  "recv measure_ctr "<<arr[i+5] << std::endl;
             }
         }
